refactor: renamed the package as it should be scalable in future developments.
