+++ adventures.and.beyond/src/input/reader/PlayerInputReader.java
@@ -1,0 +1,7 @@
+package input.reader;
+
+import input.model.player.RegisteredPlayerInput;
+
+public interface PlayerInputReader {
+    RegisteredPlayerInput readPlayerInput();
+}
