+++ adventures.and.beyond/src/input/state/ReadKeyState.java
@@ -1,4 +1,4 @@
-package keyhandler.state;
+package input.state;
 
 public interface ReadKeyState {
     boolean isUpPressed();
