+++ adventures.and.beyond/src/input/state/RegisteredKeyState.java
@@ -1,4 +1,4 @@
-package keyhandler.state;
+package input.state;
 
 public interface RegisteredKeyState extends  MutableKeyState, ReadKeyState {
 }
