--- adventures.and.beyond/src/keyhandler/playerinputstate/playerinputreader/PlayerInputReader.java
@@ -1,8 +1,0 @@
-package keyhandler.playerinputstate.playerinputreader;
-
-
-import keyhandler.playerinputstate.registeredplayerinput.RegisteredPlayerInput;
-
-public interface PlayerInputReader {
-    RegisteredPlayerInput readPlayerInput();
-}
